+++ src/main/java/org/sbrubbles/conditio/conditions/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Provides some general use conditions, which enable alternative handling protocols. The main one is
- * {@link org.sbrubbles.conditio.conditions.Notice}, intended to be the superclass of all conditions which can be left
- * unhandled without erroring out.
- */
-package org.sbrubbles.conditio.conditions;