+++ src/main/java/org/sbrubbles/conditio/policies/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides some general use policies, which enable alternative handling protocols.
+ */
+package org.sbrubbles.conditio.policies;