+++ src/test/java/org/sbrubbles/conditio/fixtures/aborting/AbortingFixture.java
@@ -4,6 +4,7 @@ import org.sbrubbles.conditio.Scope;
 import org.sbrubbles.conditio.Scopes;
 import org.sbrubbles.conditio.fixtures.BasicCondition;
 import org.sbrubbles.conditio.handlers.HandlerOps;
+import org.sbrubbles.conditio.policies.Policies;
 import org.sbrubbles.conditio.restarts.AbortException;
 import org.sbrubbles.conditio.restarts.Restarts;
 
@@ -17,7 +18,7 @@ public class AbortingFixture {
 
   public String signal() {
     try (Scope scope = Scopes.create()) {
-      scope.signal(new BasicCondition(""), Restarts.abort());
+      scope.signal(new BasicCondition(""), Policies.error(), Restarts.abort());
 
       return SIGNAL;
     } catch (Exception e) {