+++ src/test/java/org/sbrubbles/conditio/fixtures/tracing/WorkDone.java
@@ -1,8 +1,8 @@
 package org.sbrubbles.conditio.fixtures.tracing;
 
-import org.sbrubbles.conditio.conditions.Notice;
+import org.sbrubbles.conditio.Condition;
 
-public class WorkDone extends Notice {
+public class WorkDone extends Condition {
   private final int amount;
 
   public WorkDone(int amount) {